refactor: refactor core files
