--- restipie/custom_apis/__init__.py
@@ -1,1 +1,0 @@
-from . import auth, test
