--- restipie/doc_hooks/user.py
@@ -1,4 +1,0 @@
-import frappe
-
-def set_id(doc, method):
-	doc._id = frappe.generate_hash("User", 15)
