--- restipie/restipie/custom/user.json
@@ -1,61 +1,0 @@
-{
- "custom_fields": [
-  {
-   "_assign": null,
-   "_comments": null,
-   "_liked_by": null,
-   "_user_tags": null,
-   "allow_in_quick_entry": 0,
-   "allow_on_submit": 0,
-   "bold": 0,
-   "collapsible": 0,
-   "collapsible_depends_on": null,
-   "columns": 0,
-   "creation": "2021-02-17 07:44:26.898674",
-   "default": null,
-   "depends_on": null,
-   "description": null,
-   "docstatus": 0,
-   "dt": "User",
-   "fetch_from": null,
-   "fetch_if_empty": 0,
-   "fieldname": "_id",
-   "fieldtype": "Data",
-   "hidden": 0,
-   "idx": 2,
-   "ignore_user_permissions": 0,
-   "ignore_xss_filter": 0,
-   "in_global_search": 0,
-   "in_list_view": 0,
-   "in_standard_filter": 0,
-   "insert_after": "sb0_5",
-   "label": "_ID",
-   "length": 0,
-   "modified": "2021-02-17 07:44:26.898674",
-   "modified_by": "Administrator",
-   "name": "User-_id",
-   "no_copy": 0,
-   "options": null,
-   "owner": "Administrator",
-   "parent": null,
-   "parentfield": null,
-   "parenttype": null,
-   "permlevel": 0,
-   "precision": "",
-   "print_hide": 0,
-   "print_hide_if_no_value": 0,
-   "print_width": null,
-   "read_only": 1,
-   "report_hide": 0,
-   "reqd": 0,
-   "search_index": 0,
-   "translatable": 0,
-   "unique": 0,
-   "width": null
-  }
- ],
- "custom_perms": [],
- "doctype": "User",
- "property_setters": [],
- "sync_on_migrate": 1
-}
