+++ restipie/test_apis/api_version.py
@@ -1,0 +1,2 @@
+BASE_PATH = "/restipie"
+API_V1 = f"{BASE_PATH}/api/v1"
